requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
redis
pymongo
boto3
python-dotenv